+++ core/test/integration/api/api_users_spec.js
@@ -1,5 +1,5 @@
 /*globals describe, before, beforeEach, afterEach, it */
-var testUtils = require('./utils'),
+var testUtils = require('../../utils'),
     should = require('should'),
     _ = require('underscore'),
     request = require('request'),
