+++ core/test/integration/model/model_posts_spec.js
@@ -1,12 +1,12 @@
 /*globals describe, before, beforeEach, afterEach, it */
-var testUtils = require('../unit/utils'),
+var testUtils = require('../../utils'),
     should = require('should'),
     _ = require('underscore'),
     when = require('when'),
     sequence = require('when/sequence'),
 
     // Stuff we are testing
-    Models = require('../../server/models');
+    Models = require('../../../server/models');
 
 describe('Post Model', function () {
