+++ core/test/unit/client_ghostdown_spec.js
@@ -6,7 +6,7 @@
  */
 
 /*globals describe, it */
-var testUtils = require('./utils'),
+var testUtils = require('../utils'),
     should = require('should'),
 
     // Stuff we are testing
