+++ core/test/unit/import_spec.js
@@ -1,5 +1,5 @@
 /*globals describe, beforeEach, it*/
-var testUtils = require('./utils'),
+var testUtils = require('../utils'),
     should = require('should'),
     sinon = require('sinon'),
     when = require('when'),
