+++ core/test/unit/shared_gfm_spec.js
@@ -4,7 +4,7 @@
  */
 
 /*globals describe, it */
-var testUtils = require('./utils'),
+var testUtils = require('../utils'),
     should = require('should'),
 
     // Stuff we are testing
